zimkag_webapp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1

# ML / NLP
torch>=2.2.0
transformers==4.45.2
sentence-transformers==3.2.0
scikit-learn==1.5.2
numpy>=1.26
pandas>=2.2

# Knowledge graph
networkx==3.3
rapidfuzz==3.10.0

# Document parsing
pdfplumber==0.11.4
python-docx==1.1.2

# LLM client
requests==2.32.3

# PDF report
fpdf2==2.7.9
